requirements.txt
flask
flask-cors
pyaml
